Update terraform-docs settings and checkout ref
Update the default settings for terraform-docs, emptying the settings
for output-method and template, otherwise these are overridden on the
command-line by the Action and the .terraform-docs.yaml configuration
file provided is ignored.

Also update the checkout Action with the reference for the checkout so
that the git-push for terraform-docs will work again.